Updating Mechanics for Edens Gate
Sets most to generic because of platforms.  DungeonAssist and Sidestep fight where Sidestep makes you fall off so it's disabled for the fight.  We then turned off avoidance manager as well to prevent fighting from sidestep instructions.  We then set more mechanic detection and better timings to stack on a party member.  Markers and stacks are still something TBA